Split client-side sampling into estimate-local
`substreams estimate` only talks to the endpoint, `estimate-local` keeps
the streaming sampler; each command carries just the flags it uses.
